cantran/pyproject.toml
[project]
name = "cantran"
version = "0.1.0"
description = "Offline video translation to Cantonese"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "click>=8.0",
    "sounddevice>=0.4",
    "soundfile>=0.12",
    "mlx>=0.20",
    "mlx-whisper>=0.4",
    "ctranslate2>=4.0",
    "sentencepiece>=0.2",
    "opencc-python-reimplemented>=0.1.7",
    "mlx-audio>=0.1.0",
    "rich>=13.0",
    "huggingface-hub>=0.20",
    "numpy>=1.26",
    "tomli>=2.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
cantran = "cantran.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cantran"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
